Interface for the C++ spread code - mainly for internal use

raw_spread() takes a spread value, the focal cell coordinates, and a resistance matrix (all values >= 1); executes a resistant spread; and returns the final snapshot of the spread as a matrix.

Usage

raw_spread(x, spread_value, row, col, square = FALSE, symmetrical = TRUE)

Arguments

x

the resistance matrix, all values must be >= 1

spread_value

the initial bank account for the spread, which is depleted based on the resistance of each cell as the spread progresses.

row

the focal row of the spread

col

the focal column of the spread

square

if TRUE then the diagonal cells will be considered to be one cell length away from the center and the spread will tend to be square. The default of FALSE uses a diagonal length of 1.4 cells and spreads tend to be octagonal.

symmetrical

If TRUE then the average resistance of the source and neighboring destination cell is used to represent the cost of stepping to a neighboring cell. If FALSE then the destination cell is used to represent the cost of the step. The original 2009 algorithm is equivalent to FALSE. If TRUE then the cost difference between a source and any (not necessarily adjacent) destination will be the same as the cost distance from that destination to the source, and both will include half the cost of traversing the source and destination cells. If FALSE that is not guaranteed as the source cell cost is not included and the full destination cell cost is.

Value

A matrix with the same dimensions as x the values of which represent the functional proximity to the focal cell. They range from spread_value (at the focal cell) to zero for cells greater than spread_value (cost) away from the focal cell.

Details

raw_spread() calculates the functional proximity of each cell to the focal cell. With the proximity highest at the focal cell where it is equal to the spread_value and diminishing with distance from the focal cell (based on the resistance of intervening cells). The functional proximity thus is zero for cells with a functional distance greater than spread_value from the focal cell, regardless of how much greater. In practice larger spread values allow calculation of a larger functional distances but requires more processor time.

See also

Examples

# Create (minimally resistant) test matrix 
a <- matrix(1, 5, 5)
a
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    1    1    1    1
#> [2,]    1    1    1    1    1
#> [3,]    1    1    1    1    1
#> [4,]    1    1    1    1    1
#> [5,]    1    1    1    1    1
raw_spread(a, 10, 3, 3, FALSE, TRUE)
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  7.2  7.6    8  7.6  7.2
#> [2,]  7.6  8.6    9  8.6  7.6
#> [3,]  8.0  9.0   10  9.0  8.0
#> [4,]  7.6  8.6    9  8.6  7.6
#> [5,]  7.2  7.6    8  7.6  7.2

# Add two higher resistance bands
a[2, ]  <- 5
a[,2] <- 5
a
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    5    1    1    1
#> [2,]    5    5    5    5    5
#> [3,]    1    5    1    1    1
#> [4,]    1    5    1    1    1
#> [5,]    1    5    1    1    1
# Average of source and destination resistance used with each step
raw_spread(a, 10, 3, 3, square = FALSE, symmetrical =  FALSE) 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  1.6    0    4  3.6  2.6
#> [2,]  0.0    3    5  4.0  3.0
#> [3,]  4.0    5   10  9.0  8.0
#> [4,]  3.6    4    9  8.6  7.6
#> [5,]  2.6    3    8  7.6  7.2

# Destination resistance used with each step
raw_spread(a, 10, 3, 3, square = FALSE, symmetrical = TRUE)  
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  1.6  1.0    4  3.0  2.0
#> [2,]  1.0  5.8    7  6.0  5.0
#> [3,]  4.0  7.0   10  9.0  8.0
#> [4,]  3.0  6.0    9  8.6  7.6
#> [5,]  2.0  5.0    8  7.6  7.2

# Square: the diagonal cost same as orthogonal 
a[,] <- 1
raw_spread(a, 10, 3, 3, square = TRUE, symmetrical =  FALSE) 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    8    8    8    8    8
#> [2,]    8    9    9    9    8
#> [3,]    8    9   10    9    8
#> [4,]    8    9    9    9    8
#> [5,]    8    8    8    8    8